fix: replicate #2183 for new client
Avoid loading all notes in memory on input creation.
