+++ fedimint-client-legacy/src/mint/mod.rs
@@ -1,6 +1,5 @@
 pub mod db;
 
-use std::cmp::Ordering;
 use std::fmt;
 use std::pin::Pin;
 use std::sync::Arc;
@@ -16,7 +15,7 @@ use fedimint_core::module::registry::ModuleDecoderRegistry;
 use fedimint_core::module::{ModuleCommon, TransactionItemAmount};
 use fedimint_core::tiered::InvalidAmountTierError;
 use fedimint_core::{Amount, OutPoint, Tiered, TieredMulti, TieredSummary, TransactionId};
-use fedimint_mint_client::MintModuleTypes;
+use fedimint_mint_client::{select_notes_from_stream, MintModuleTypes};
 use futures::executor::block_on;
 use futures::{Future, StreamExt};
 use secp256k1_zkp::{KeyPair, Secp256k1, Signing};
@@ -411,75 +410,11 @@ impl MintClient {
             .find_by_prefix_sorted_descending(&NoteKeyPrefix)
             .await
             .map(|(key, note)| (key.amount, note));
-        Self::select_notes_from_stream(note_stream, amount).await
-    }
-
-    // We are using a greedy algorithm to select notes. We start with the largest
-    // then proceed to the lowest tiers/denominations.
-    // But there is a catch: we don't know if there are enough notes in the lowest
-    // tiers, so we need to save a big note in case the sum of the following
-    // small notes are not enough.
-    async fn select_notes_from_stream<Note>(
-        stream: impl futures::Stream<Item = (Amount, Note)>,
-        requested_amount: Amount,
-    ) -> Result<TieredMulti<Note>> {
-        if requested_amount == Amount::ZERO {
-            return Ok(TieredMulti::default());
-        }
-        let mut stream = Box::pin(stream);
-        let mut selected = vec![];
-        // This is the big note we save in case the sum of the following small notes are
-        // not sufficient to cover the pending amount
-        // The tuple is (amount, note, checkpoint), where checkpoint is the index where
-        // the note should be inserted on the selected vector if it is needed
-        let mut last_big_note_checkpoint: Option<(Amount, Note, usize)> = None;
-        let mut pending_amount = requested_amount;
-        let mut previous_amount: Option<Amount> = None; // used to assert descending order
-        loop {
-            if let Some((note_amount, note)) = stream.next().await {
-                assert!(
-                    previous_amount.map_or(true, |previous| previous >= note_amount),
-                    "notes are not sorted in descending order"
-                );
-                previous_amount = Some(note_amount);
-                match note_amount.cmp(&pending_amount) {
-                    Ordering::Less => {
-                        // keep adding notes until we have enough
-                        pending_amount -= note_amount;
-                        selected.push((note_amount, note))
-                    }
-                    Ordering::Greater => {
-                        // probably we don't need this big note, but we'll keep it in case the
-                        // following small notes don't add up to the
-                        // requested amount
-                        last_big_note_checkpoint = Some((note_amount, note, selected.len()));
-                    }
-                    Ordering::Equal => {
-                        // exactly enough notes, return
-                        selected.push((note_amount, note));
-                        return Ok(selected.into_iter().collect());
-                    }
-                }
-            } else {
-                assert!(pending_amount > Amount::ZERO);
-                if let Some((big_note_amount, big_note, checkpoint)) = last_big_note_checkpoint {
-                    // the sum of the small notes don't add up to the pending amount, remove
-                    // them
-                    selected.truncate(checkpoint);
-                    // and use the big note to cover it
-                    selected.push((big_note_amount, big_note));
-                    // so now we have enough to cover the requested amount, return
-                    return Ok(selected.into_iter().collect());
-                } else {
-                    let total_amount = requested_amount - pending_amount;
-                    // not enough notes, return
-                    return Err(MintClientError::InsufficientBalance(
-                        requested_amount,
-                        total_amount,
-                    ));
-                }
-            }
-        }
+        select_notes_from_stream(note_stream, amount)
+            .await
+            .map_err(|err| {
+                MintClientError::InsufficientBalance(err.requested_amount, err.total_amount)
+            })
     }
 
     pub async fn receive_notes(
@@ -756,11 +691,10 @@ mod tests {
     use fedimint_core::db::Database;
     use fedimint_core::module::registry::ModuleDecoderRegistry;
     use fedimint_core::outcome::{SerdeOutputOutcome, TransactionStatus};
-    use fedimint_core::{Amount, OutPoint, ServerModule, Tiered, TieredMulti, TransactionId};
+    use fedimint_core::{Amount, OutPoint, ServerModule, Tiered, TransactionId};
     use fedimint_mint_server::{Mint, MintGen, MintGenParams};
     use fedimint_testing::FakeFed;
     use futures::executor::block_on;
-    use itertools::Itertools;
     use tokio::sync::Mutex;
 
     use super::*;
@@ -1113,110 +1047,5 @@ mod tests {
             .unwrap_or(0);
         // Ensure we didn't skip any keys
         assert_eq!(last_idx, result_count as u64);
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn select_notes_avg_test() {
-        let max_amount = Amount::from_sats(1000000);
-        let tiers = Tiered::gen_denominations(max_amount);
-        let tiered =
-            TieredSummary::represent_amount::<()>(max_amount, &Default::default(), &tiers, 3);
-
-        let mut total_notes = 0;
-        for multiplier in 1..100 {
-            let stream = reverse_sorted_note_stream(tiered.iter().collect());
-            let select =
-                MintClient::select_notes_from_stream(stream, Amount::from_sats(multiplier * 1000))
-                    .await;
-            total_notes += select.unwrap().into_iter_items().count();
-        }
-        assert_eq!(total_notes / 100, 10);
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn select_notes_returns_exact_amount_with_minimum_notes() {
-        let f = || {
-            reverse_sorted_note_stream(vec![
-                (Amount::from_sats(1), 10),
-                (Amount::from_sats(5), 10),
-                (Amount::from_sats(20), 10),
-            ])
-        };
-        assert_eq!(
-            MintClient::select_notes_from_stream(f(), Amount::from_sats(7))
-                .await
-                .unwrap(),
-            notes(vec![(Amount::from_sats(1), 2), (Amount::from_sats(5), 1)])
-        );
-        assert_eq!(
-            MintClient::select_notes_from_stream(f(), Amount::from_sats(20))
-                .await
-                .unwrap(),
-            notes(vec![(Amount::from_sats(20), 1)])
-        );
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn select_notes_returns_next_smallest_amount_if_exact_change_cannot_be_made() {
-        let stream = reverse_sorted_note_stream(vec![
-            (Amount::from_sats(1), 1),
-            (Amount::from_sats(5), 5),
-            (Amount::from_sats(20), 5),
-        ]);
-        assert_eq!(
-            MintClient::select_notes_from_stream(stream, Amount::from_sats(7))
-                .await
-                .unwrap(),
-            notes(vec![(Amount::from_sats(5), 2)])
-        );
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn select_notes_uses_big_note_if_small_amounts_are_not_sufficient() {
-        let stream = reverse_sorted_note_stream(vec![
-            (Amount::from_sats(1), 3),
-            (Amount::from_sats(5), 3),
-            (Amount::from_sats(20), 2),
-        ]);
-        assert_eq!(
-            MintClient::select_notes_from_stream(stream, Amount::from_sats(39))
-                .await
-                .unwrap(),
-            notes(vec![(Amount::from_sats(20), 2)])
-        );
-    }
-
-    #[test_log::test(tokio::test)]
-    async fn select_notes_returns_error_if_amount_is_too_large() {
-        let stream = reverse_sorted_note_stream(vec![(Amount::from_sats(10), 1)]);
-        match MintClient::select_notes_from_stream(stream, Amount::from_sats(100))
-            .await
-            .unwrap_err()
-        {
-            MintClientError::InsufficientBalance(_, total) => {
-                assert_eq!(total, Amount::from_sats(10))
-            }
-            other => panic!("Unexpected error: {other:?}"),
-        };
-    }
-
-    fn reverse_sorted_note_stream(
-        notes: Vec<(Amount, usize)>,
-    ) -> impl futures::Stream<Item = (Amount, String)> {
-        futures::stream::iter(
-            notes
-                .into_iter()
-                // We are creating `number` dummy notes of `amount` value
-                .flat_map(|(amount, number)| vec![(amount, "dummy note".into()); number])
-                .sorted()
-                .rev(),
-        )
-    }
-
-    fn notes(notes: Vec<(Amount, usize)>) -> TieredMulti<String> {
-        notes
-            .into_iter()
-            .flat_map(|(amount, number)| vec![(amount, "dummy note".into()); number])
-            .collect()
     }
 }
