+++ modules/fedimint-mint-client/Cargo.toml
@@ -38,3 +38,4 @@ tracing ="0.1.37"
 rand = "0.8"
 tracing-subscriber = { version = "0.3.16", features = [ "env-filter" ] }
 test-log = { version = "0.2", features = [ "trace" ], default-features = false }
+tokio = { version = "1.27.0", features = [ "macros", "rt" ] }
